[feature] Extract FeatureExtractor class from main_featureExtraction.cpp
This allows the class to be reused by other components.
