+++ src/aliceVision/feature/CMakeLists.txt
@@ -14,6 +14,7 @@ set(features_files_headers
   sift/SIFT.hpp
   Descriptor.hpp
   feature.hpp
+  FeatureExtractor.hpp
   FeaturesPerView.hpp
   Hamming.hpp
   ImageDescriber.hpp
@@ -34,6 +35,7 @@ set(features_files_sources
   akaze/ImageDescriber_AKAZE.cpp
   sift/SIFT.cpp
   sift/ImageDescriber_DSPSIFT_vlfeat.cpp
+  FeatureExtractor.cpp
   FeaturesPerView.cpp
   ImageDescriber.cpp
   imageDescriberCommon.cpp
