+++ src/software/pipeline/main_featureExtraction.cpp
@@ -5,7 +5,7 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <aliceVision/alicevision_omp.hpp>
+#include <aliceVision/feature/FeatureExtractor.hpp>
 #include <aliceVision/sfmData/SfMData.hpp>
 #include <aliceVision/sfmDataIO/sfmDataIO.hpp>
 #include <aliceVision/feature/imageDescriberCommon.hpp>
@@ -15,8 +15,6 @@
 #define ALICEVISION_HAVE_GPU_FEATURES
 #include <aliceVision/gpu/gpu.hpp>
 #endif
-#include <aliceVision/image/all.hpp>
-#include <aliceVision/system/MemoryInfo.hpp>
 #include <aliceVision/system/Timer.hpp>
 #include <aliceVision/system/Logger.hpp>
 #include <aliceVision/system/main.hpp>
@@ -27,13 +25,9 @@
 #include <boost/filesystem.hpp>
 
 #include <string>
-#include <fstream>
-#include <sstream>
 #include <iostream>
 #include <functional>
 #include <memory>
-#include <limits>
-#include <iomanip>
 
 // These constants define the current software version.
 // They must be updated when the command line is changed.
@@ -44,298 +38,6 @@ using namespace aliceVision;
 
 namespace po = boost::program_options;
 namespace fs = boost::filesystem;
-
-class FeatureExtractor
-{
-  struct ViewJob
-  {
-    const sfmData::View& view;
-    std::size_t memoryConsuption = 0;
-    std::string outputBasename;
-    std::vector<std::size_t> cpuImageDescriberIndexes;
-    std::vector<std::size_t> gpuImageDescriberIndexes;
-
-    ViewJob(const sfmData::View& view,
-            const std::string& outputFolder)
-      : view(view)
-      , outputBasename(fs::path(fs::path(outputFolder) / fs::path(std::to_string(view.getViewId()))).string())
-    {}
-
-    ~ViewJob() = default;
-
-    bool useGPU() const
-    {
-      return !gpuImageDescriberIndexes.empty();
-    }
-
-    bool useCPU() const
-    {
-      return !cpuImageDescriberIndexes.empty();
-    }
-
-    std::string getFeaturesPath(feature::EImageDescriberType imageDescriberType) const
-    {
-      return outputBasename + "." + feature::EImageDescriberType_enumToString(imageDescriberType) + ".feat";
-    }
-
-    std::string getDescriptorPath(feature::EImageDescriberType imageDescriberType) const
-    {
-      return outputBasename + "." + feature::EImageDescriberType_enumToString(imageDescriberType) + ".desc";
-    }
-
-    void setImageDescribers(const std::vector<std::shared_ptr<feature::ImageDescriber>>& imageDescribers)
-    {
-      for(std::size_t i = 0; i < imageDescribers.size(); ++i)
-      {
-        const std::shared_ptr<feature::ImageDescriber>& imageDescriber = imageDescribers.at(i);
-        feature::EImageDescriberType imageDescriberType = imageDescriber->getDescriberType();
-
-        if(fs::exists(getFeaturesPath(imageDescriberType)) &&
-           fs::exists(getDescriptorPath(imageDescriberType)))
-          continue;
-
-        memoryConsuption += imageDescriber->getMemoryConsumption(view.getWidth(), view.getHeight());
-
-        if(imageDescriber->useCuda())
-          gpuImageDescriberIndexes.push_back(i);
-        else
-          cpuImageDescriberIndexes.push_back(i);
-      }
-    }
-  };
-
-public:
-
-  explicit FeatureExtractor(const sfmData::SfMData& sfmData)
-    : _sfmData(sfmData)
-  {}
-
-  void setRange(int rangeStart, int rangeSize)
-  {
-    _rangeStart = rangeStart;
-    _rangeSize = rangeSize;
-  }
-
-  void setMaxThreads(int maxThreads)
-  {
-    _maxThreads = maxThreads;
-  }
-
-  void setMasksFolder(const std::string& folder)
-  {
-    _masksFolder = folder;
-  }
-
-  void setOutputFolder(const std::string& folder)
-  {
-    _outputFolder = folder;
-  }
-
-  void addImageDescriber(std::shared_ptr<feature::ImageDescriber>& imageDescriber)
-  {
-    _imageDescribers.push_back(imageDescriber);
-  }
-
-  void process()
-  {
-    // iteration on each view in the range in order
-    // to prepare viewJob stack
-    sfmData::Views::const_iterator itViewBegin = _sfmData.getViews().begin();
-    sfmData::Views::const_iterator itViewEnd = _sfmData.getViews().end();
-
-    if(_rangeStart != -1)
-    {
-      std::advance(itViewBegin, _rangeStart);
-      itViewEnd = itViewBegin;
-      std::advance(itViewEnd, _rangeSize);
-    }
-
-    std::size_t jobMaxMemoryConsuption = 0;
-
-    for(auto it = itViewBegin; it != itViewEnd; ++it)
-    {
-      const sfmData::View& view = *(it->second.get());
-      ViewJob viewJob(view, _outputFolder);
-
-      viewJob.setImageDescribers(_imageDescribers);
-      jobMaxMemoryConsuption = std::max(jobMaxMemoryConsuption, viewJob.memoryConsuption);
-
-      if(viewJob.useCPU())
-        _cpuJobs.push_back(viewJob);
-
-      if(viewJob.useGPU())
-        _gpuJobs.push_back(viewJob);
-    }
-
-    if(!_cpuJobs.empty())
-    {
-      system::MemoryInfo memoryInformation = system::getMemoryInfo();
-
-      ALICEVISION_LOG_INFO("Job max memory consumption for one image: " << jobMaxMemoryConsuption / (1024*1024) << " MB");
-      ALICEVISION_LOG_INFO("Memory information: " << std::endl << memoryInformation);
-
-      if(jobMaxMemoryConsuption == 0)
-        throw std::runtime_error("Cannot compute feature extraction job max memory consumption.");
-
-      // How many buffers can fit in 90% of the available RAM?
-      // This is used to estimate how many jobs can be computed in parallel without SWAP.
-      const std::size_t memoryImageCapacity = std::size_t((0.9 * memoryInformation.availableRam) / jobMaxMemoryConsuption);
-      std::size_t nbThreads = std::max(std::size_t(1), memoryImageCapacity);
-      ALICEVISION_LOG_INFO("Max number of threads regarding memory usage: " << nbThreads);
-      const double oneGB = 1024.0 * 1024.0 * 1024.0;
-      if(jobMaxMemoryConsuption > memoryInformation.availableRam)
-      {
-          ALICEVISION_LOG_WARNING("The amount of RAM available is critical to extract features.");
-          if(jobMaxMemoryConsuption <= memoryInformation.totalRam)
-          {
-              ALICEVISION_LOG_WARNING("But the total amount of RAM is enough to extract features, so you should close other running applications.");
-              ALICEVISION_LOG_WARNING(" => " << std::size_t(std::round((double(memoryInformation.totalRam - memoryInformation.availableRam) / oneGB)))
-                                      << " GB are used by other applications for a total RAM capacity of "
-                                      << std::size_t(std::round(double(memoryInformation.totalRam) / oneGB)) << " GB.");
-          }
-      }
-      else
-      {
-          if(memoryInformation.availableRam < 0.5 * memoryInformation.totalRam)
-          {
-              ALICEVISION_LOG_WARNING("More than half of the RAM is used by other applications. It would be more efficient to close them.");
-              ALICEVISION_LOG_WARNING(" => "
-                                      << std::size_t(std::round(double(memoryInformation.totalRam - memoryInformation.availableRam) / oneGB))
-                                      << " GB are used by other applications for a total RAM capacity of "
-                                      << std::size_t(std::round(double(memoryInformation.totalRam) / oneGB)) << " GB.");
-          }
-      }
-
-      if(memoryInformation.availableRam == 0)
-      {
-        ALICEVISION_LOG_WARNING("Cannot find available system memory, this can be due to OS limitation.\n"
-                                "Use only one thread for CPU feature extraction.");
-        nbThreads = 1;
-      }
-
-      // nbThreads should not be higher than user maxThreads param
-      if(_maxThreads > 0)
-        nbThreads = std::min(static_cast<std::size_t>(_maxThreads), nbThreads);
-
-      // nbThreads should not be higher than the core number
-      nbThreads = std::min(static_cast<std::size_t>(omp_get_num_procs()), nbThreads);
-
-      // nbThreads should not be higher than the number of jobs
-      nbThreads = std::min(_cpuJobs.size(), nbThreads);
-
-      ALICEVISION_LOG_INFO("# threads for extraction: " << nbThreads);
-      omp_set_nested(1);
-
-#pragma omp parallel for num_threads(nbThreads)
-      for(int i = 0; i < _cpuJobs.size(); ++i)
-        computeViewJob(_cpuJobs.at(i));
-    }
-
-    if(!_gpuJobs.empty())
-    {
-      for(const auto& job : _gpuJobs)
-        computeViewJob(job, true);
-    }
-  }
-
-  ~FeatureExtractor() = default;
-
-private:
-
-  void computeViewJob(const ViewJob& job, bool useGPU = false)
-  {
-    image::Image<float> imageGrayFloat;
-    image::Image<unsigned char> imageGrayUChar;
-    image::Image<unsigned char> mask;
-
-    image::readImage(job.view.getImagePath(), imageGrayFloat, image::EImageColorSpace::SRGB);
-
-    if(!_masksFolder.empty() && fs::exists(_masksFolder))
-    {
-      const auto masksFolder = fs::path(_masksFolder);
-      const auto idMaskPath = masksFolder / fs::path(std::to_string(job.view.getViewId())).replace_extension("png");
-      const auto nameMaskPath = masksFolder / fs::path(job.view.getImagePath()).filename().replace_extension("png");
-
-      if(fs::exists(idMaskPath))
-      {
-        image::readImage(idMaskPath.string(), mask, image::EImageColorSpace::LINEAR);
-      }
-      else if(fs::exists(nameMaskPath))
-      {
-        image::readImage(nameMaskPath.string(), mask, image::EImageColorSpace::LINEAR);
-      }
-    }
-
-    const auto imageDescriberIndexes = useGPU ? job.gpuImageDescriberIndexes : job.cpuImageDescriberIndexes;
-
-    for(const auto & imageDescriberIndex : imageDescriberIndexes)
-    {
-      const auto& imageDescriber = _imageDescribers.at(imageDescriberIndex);
-      const feature::EImageDescriberType imageDescriberType = imageDescriber->getDescriberType();
-      const std::string imageDescriberTypeName = feature::EImageDescriberType_enumToString(imageDescriberType);
-
-      // Compute features and descriptors and export them to files
-      ALICEVISION_LOG_INFO("Extracting " << imageDescriberTypeName  << " features from view '" << job.view.getImagePath() << "' " << (useGPU ? "[gpu]" : "[cpu]"));
-
-      std::unique_ptr<feature::Regions> regions;
-      if(imageDescriber->useFloatImage())
-      {
-        // image buffer use float image, use the read buffer
-        imageDescriber->describe(imageGrayFloat, regions);
-      }
-      else
-      {
-        // image buffer can't use float image
-        if(imageGrayUChar.Width() == 0) // the first time, convert the float buffer to uchar
-          imageGrayUChar = (imageGrayFloat.GetMat() * 255.f).cast<unsigned char>();
-        imageDescriber->describe(imageGrayUChar, regions);
-      }
-
-      if(mask.Height() > 0)
-      {
-        std::vector<feature::FeatureInImage> selectedIndices;
-        for(size_t i=0, n=regions->RegionCount(); i != n; ++i)
-        {
-          const Vec2 position = regions->GetRegionPosition(i);
-          const int x = int(position.x());
-          const int y = int(position.y());
-
-          bool masked = false;
-          if(x < mask.Width() && y < mask.Height())
-          {
-            if(mask(y, x) == 0)
-            {
-              masked = true;
-            }
-          }
-
-          if(!masked)
-          {
-            selectedIndices.push_back({IndexT(i), 0});
-          }
-        }
-
-        std::vector<IndexT> out_associated3dPoint;
-        std::map<IndexT, IndexT> out_mapFullToLocal;
-        regions = regions->createFilteredRegions(selectedIndices, out_associated3dPoint, out_mapFullToLocal);
-      }
-
-      imageDescriber->Save(regions.get(), job.getFeaturesPath(imageDescriberType), job.getDescriptorPath(imageDescriberType));
-      ALICEVISION_LOG_INFO(std::left << std::setw(6) << " " << regions->RegionCount() << " " << imageDescriberTypeName  << " features extracted from view '" << job.view.getImagePath() << "'");
-    }
-  }
-
-  const sfmData::SfMData& _sfmData;
-  std::vector<std::shared_ptr<feature::ImageDescriber>> _imageDescribers;
-  std::string _masksFolder;
-  std::string _outputFolder;
-  int _rangeStart = -1;
-  int _rangeSize = -1;
-  int _maxThreads = -1;
-  std::vector<ViewJob> _cpuJobs;
-  std::vector<ViewJob> _gpuJobs;
-};
-
 
 /// - Compute view image description (feature & descriptor extraction)
 /// - Export computed data
@@ -463,7 +165,7 @@ int aliceVision_main(int argc, char **argv)
   }
 
   // create feature extractor
-  FeatureExtractor extractor(sfmData);
+  feature::FeatureExtractor extractor(sfmData);
   extractor.setMasksFolder(masksFolder);
   extractor.setOutputFolder(outputFolder);
